Remove one warning

+++ src/opengl/textureunit.cpp
@@ -86,7 +86,7 @@ void TextureUnit::setZeroUnit() {
 }
 
 void TextureUnit::deinitialize() {
-	for (int i = 0; i < _busyUnits.size(); ++i)
+	for (size_t i = 0; i < _busyUnits.size(); ++i)
 		_busyUnits[i] = false;
     _totalActive = 0;
     setZeroUnit();
